% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_warping_mixture.R
\name{base_shape}
\alias{base_shape}
\title{Make base shape by standarding a symmetric beta-density}
\usage{
base_shape(x, a = 2)
}
\arguments{
\item{x}{numeric vector between 0 and 1 where the function is evaluated}

\item{a}{non-negative parameters of the Beta distribution}
}
\description{
Internal function to be called by sim_warping_mixture()
}
