% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_saem.R
\name{control_saem}
\alias{control_saem}
\title{Auxiliary for Specifying the control parameters for the stochastic approximation}
\usage{
control_saem(n_saem_iter = 1, n_saem_burn = 1, n_mcmc_burn = 1,
  n_core = 1, saem_step_seq_pow = 1, prop_sigma = 0.01,
  need_centering = FALSE, accept_rate_lb = 0.17, accept_rate_ub = 0.33,
  accept_rate_window = 5, f_order = 4, h_order = 4, f_knots = NULL,
  h_knots = NULL, f_n_knots = NULL, h_n_knots = NULL, ind_amp = FALSE,
  seed = NULL)
}
\arguments{
\item{n_saem_iter}{Number of EM iterations (in addition to SAEM burn)}

\item{n_saem_burn}{Number of SAEM burn-in steps}

\item{n_mcmc_burn}{Number of MCMC steps}

\item{n_core}{Number of cores for parallel computing}

\item{saem_step_seq_pow}{Power of the harmonic series of SAEM step sizes}

\item{prop_sigma}{Initial value for the variance of the MH proposal}

\item{need_centering}{Logical variable of whether the sampled amplitude effects should be centered to have sample mean equals to mu}

\item{accept_rate_lb}{Lower bound of the desired MH acceptance rate}

\item{accept_rate_ub}{Upper bound of the desired MH acceptance rate}

\item{accept_rate_window}{Number of E-steps between successive calibration of MH proposal}

\item{f_order}{Order of the spline for the base shape}

\item{h_order}{Order of the splines for the warping functions}

\item{f_knots}{Knot locations of the base shape spline (including boundary knots)}

\item{h_knots}{Knot locations of the warping function splines (including boundary knots)}

\item{f_n_knots}{Number of equally spaced knots, from 0 to 1, to use if f_knots is not specified}

\item{h_n_knots}{Number of equally spaced knots, from 0 to 1, to use if h_knots is not specified}
}
\description{
This function help specify the control parameters for the stochastic approximation.
}
