% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mixed_shape.R
\name{fit_mixed_shape}
\alias{fit_mixed_shape}
\title{Fit the mixture model of Brillinger and Stewart (1997)}
\usage{
fit_mixed_shape(y, x = NULL, nCat = 9, nIter = 100, plot.est = FALSE)
}
\arguments{
\item{y}{matrix, numeric matrix containing depth. Each column represents one dive.}

\item{x}{numeric, common time grid at which the depths are recorded}

\item{nCat}{integer, number of dive categories}

\item{nIter}{integer, number of maximum EM iterations}

\item{plot.est}{if TRUE, plot estimated shapes}
}
\description{
This function fits the mixture of shape model using the EM algorithm
described in Brillinger and Stewart (1997).
}
