% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsim_mixture_a_model.R
\name{fsim_mixture_a_model}
\alias{fsim_mixture_a_model}
\title{Fitting shape-invariant model with mixture of shape and amplitude shifting and scaling
(Obsolete and probably buggy)}
\usage{
fsim_mixture_a_model(y, obs_time, curve_id, init_clust = NULL, n_clust = 1,
  saem_control = control_saem(), trace = FALSE)
}
\arguments{
\item{y}{a vector of observed curves}

\item{obs_time}{a vector of the observation times}

\item{curve_id}{a vector of curve IDs}

\item{saem_control}{a list of values to control the MCMC and stochastic approximation}
}
\description{
This function fits the model \deqn{Y_i(t) = a_{i,sh} + a_{i,sc} f_m(t)}
by maximum likelihood via a stochastic approximation EM algorithm. In the model,
$f_m$ is a B-spline representing a common shape for cluster $m$.
}
