% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsim_wllk.R
\name{fsim_wllk}
\alias{fsim_wllk}
\title{Function to compute the mixture of warping likelihood}
\usage{
fsim_wllk(fsim_obj)
}
\arguments{
\item{fsim_obj}{: object returned by the mixture of warping model with fixed unimodel template}
}
\description{
This function computes the likelihood of the warping component
for the mixture of warping function. The predicted warping
coefficients by stochastic approximation is treated as the data;
the likelihood of the mixture of Dirichlet distributions is
returned
}
