% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik_fsim.R
\name{logLik_fsim}
\alias{logLik_fsim}
\title{Observed data log-likelihood of the registration model}
\usage{
logLik_fsim(saemObj, method = c("joint_mc", "joint_is", "marginal_is",
  "mcmc"), n_sim = 10, seed = 0)
}
\arguments{
\item{saemObj}{fitted registration model}

\item{method}{method for computing the log-likelihood}

\item{n_sim}{number of Monte Carlo sample}

\item{seed}{seed for random number generator}
}
\description{
Compute log-likelihood of the fitted model by sampling based approximation
}
