% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_shape_wllk.R
\name{mixed_shape_wllk}
\alias{mixed_shape_wllk}
\title{Function to compute the mixture of warping likelihood}
\usage{
mixed_shape_wllk(y, bs97_obj)
}
\arguments{
\item{y}{: n x p matrix of the n curves}

\item{bs97_obj}{: object returned by mixed_shape that fits Brillinger's model}
}
\description{
This function computes the likelihood of the warping component
for the mixture of warping function. The predicted warping
coefficients by stochastic approximation is treated as the data;
the likelihood of the mixture of Dirichlet distributions is
returned
}
