% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_of_dirichlet.R
\name{mixture_of_dirichlet}
\alias{mixture_of_dirichlet}
\title{Fit finite mixture of Dirichlet by EM algorithm}
\usage{
mixture_of_dirichlet(w, nclust, init_clust = NULL, maxit = 20, nstart = 1)
}
\arguments{
\item{w}{data}

\item{nclust}{number of mixture component}

\item{init_clust}{vector of initial clustering label}

\item{maxit}{maximum number of iterations for EM algorithm}

\item{nstart}{number of random start if init_clust is not specified}
}
\description{
Fit finite mixture of Dirichlet by EM algorithm
}
