% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_of_dirichlet_logLik.R
\name{mixture_of_dirichlet_logLik}
\alias{mixture_of_dirichlet_logLik}
\title{Function to compute the log-likelihood, AIC and BIC for the
mixture of dirichlet distribution}
\usage{
mixture_of_dirichlet_logLik(dw_mat, kappa_hat, p_hat)
}
\arguments{
\item{dw_mat}{k x n matrix of k-dimensional vector of compositional data}

\item{kappa_hat}{k x M matrix of shape parameters for the M clusters}

\item{p_hat}{M x 1 vector of mixing proportion}
}
\description{
Function to compute the log-likelihood, AIC and BIC for the
mixture of dirichlet distribution
}
