% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_of_dirichlet_wllk.R
\name{mixture_of_dirichlet_wllk}
\alias{mixture_of_dirichlet_wllk}
\title{Function to compute the mixture of warping likelihood}
\usage{
mixture_of_dirichlet_wllk(md_obj)
}
\arguments{
\item{md_obj}{: object returned by mixture_of_dirichlet()}
}
\description{
This function computes the likelihood of the warping component
for the mixture of warping function. The predicted warping
coefficients by stochastic approximation is treated as the data;
the likelihood of the mixture of Dirichlet distributions is
returned
}
