% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{saem_fit_mixed_warping}
\alias{saem_fit_mixed_warping}
\title{Internal function for fitting the model by SAEM}
\usage{
saem_fit_mixed_warping(curve_list, pars_list, control_list, y_scaling_factor,
  trace)
}
\arguments{
\item{curve_list}{List of curve objects, each should have a component called data, with subcomponents, x and y,
for the time and value of the functional data}

\item{pars_list}{List with components: mu, kappa and sigma2}

\item{control_list}{List generated by control_saem()}

\item{y_scaling_factor}{Numeric value for scaling back the estimates and predictions to the original scale of y}
}
\description{
Internal function for fitting the model by SAEM
}
