% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_warping_mixture.R
\name{sim_warping_mixture}
\alias{sim_warping_mixture}
\title{Simulate warping mixture}
\usage{
sim_warping_mixture(n, p, kappa, tau = 1, ni = 1001, mu_sh = -15,
  mu_sc = 500, sd_sh = 5, sd_sc = 50, sd_err = 5)
}
\description{
Simulate warping mixture
}
