% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dirichlet_mle}
\alias{dirichlet_mle}
\title{Internal function for fitting the model by SAEM}
\usage{
dirichlet_mle(w_list)
}
\arguments{
\item{w_list}{R List - list of compositional data points}
}
\description{
Internal function for fitting the model by SAEM
}
