% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsim_mixed_shape.R
\name{fsim_mixed_shape}
\alias{fsim_mixed_shape}
\title{Fitting mixture of shapes model with amplitude shifting and scaling by SAEM}
\usage{
fsim_mixed_shape(y, obs_time, curve_id, init_clust = NULL, n_clust = 1,
  saem_control = control_saem(), trace = FALSE)
}
\arguments{
\item{y}{a vector of observed curves}

\item{obs_time}{a vector of the observation times}

\item{curve_id}{a vector of curve IDs}

\item{init_clust}{vector of inital clutering label with length equals to the number of curves}

\item{n_clust}{integer, number of clusters}

\item{saem_control}{a list of values to control the MCMC and stochastic approximation. See control_saem().}

\item{trace}{if TRUE, tracing information of the estimated parameters are printed}
}
\description{
This function fits the model \deqn{Y_i(t) = a_{i,sh} + a_{i,sc} f_{m_i}(t)}
by maximum likelihood via a stochastic approximation EM algorithm. In the model,
$f_m$ is a B-spline representing a common shape for cluster $m$. This model assume
no warpings.
}
