% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsim_mixed_warping.R
\name{fsim_mixed_warping}
\alias{fsim_mixed_warping}
\title{Fitting shape-invariant model with flexible time transformations that follows
a finite mixture of dirichlet distribution
(obsolete implementation, see fsim_mixed_warped_curves)}
\usage{
fsim_mixed_warping(y, obs_time, curve_id, init_clust = NULL, n_clust = 1,
  saem_control = control_saem(), trace = FALSE)
}
\arguments{
\item{y}{a vector of observed curves}

\item{obs_time}{a vector of the observation times}

\item{curve_id}{a vector of curve IDs}

\item{init_clust}{vector of inital clutering label with length equals to the number of curves}

\item{n_clust}{integer, number of clusters}

\item{saem_control}{a list of values to control the MCMC and stochastic approximation. See control_saem().}

\item{trace}{if TRUE, tracing information of the estimated parameters are printed}
}
\description{
This function fits the model \deqn{Y_i(t) = a_{i,sh} + a_{i,sc} f \circ h_i(y)}
by maximum likelihood via a stochastic approximation EM algorithm. In the model,
$f$ is a B-spline representing a common shape whereas \eqn{h_i:[0, 1] \to [0, 1]}
is a monotone B-spline representing a random time transformation
}
