% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mixed_shape.R
\name{mixed_shape}
\alias{mixed_shape}
\title{Fit the mixture model of Brillinger and Stewart (1997)}
\usage{
mixed_shape(y, x = NULL, nCat = 9, nIter = 100, nTry = 10)
}
\arguments{
\item{y}{matrix, numeric matrix containing depth. Each column represents one dive.}

\item{x}{numeric, common time grid at which the depths are recorded}

\item{nCat}{integer, number of dive categories}

\item{nIter}{integer, number of EM iterations}

\item{nTry}{integer, number of multi-random start}
}
\description{
This function fits the mixture of shape model using the EM algorithm
described in Brillinger and Stewart (1997).
}
