% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_warping_mixture.R
\name{sim_warping_mixture}
\alias{sim_warping_mixture}
\title{Simulate curves with mixture of warping}
\usage{
sim_warping_mixture(n, p, kappa, tau = 1, ni = 1001, mu_sh = -15,
  mu_sc = 500, sd_sh = 5, sd_sc = 50, sd_err = 5, shape_a = 2,
  shape_b = 2)
}
\arguments{
\item{n}{integer, number of curves}

\item{p}{numeric vector, mixture proportions}

\item{kappa}{G x K nuermic matrix, means of the Dirichlet distribution
for the G mixture component, stacked in rows}

\item{tau}{numeric, common concentration parameter of the Dirichlet components}

\item{ni}{numeric, number of points per curve}

\item{mu_sh}{numeric, mean of amplitude shifting effect}

\item{mu_sc}{numeric, mean of amplitude scaling effect}

\item{sd_sh}{numeric, standard deviation of amplitude shifting}

\item{sd_sc}{numeric, standard deviation of amplitude scaling}

\item{sd_err}{numeric, standard deviation of the residual term}

\item{shape_a}{numeric, non-negative parameters of the Beta distribution}

\item{shape_b}{numeric, non-negative parameters of the Beta distribution}
}
\description{
This function simulate observations from the mixture of warping model
with a unimodal base shape. The base shape is negative of a standardized beta density
function with minimum equals to -1
}
\examples{
require(ggplot2)
require(fda)
kappa0 <- c(1,2,2,1)
kappa0 <- kappa0 / sum(kappa0)
kappa1 <- c(0,4,4,0)
kappa1 <- kappa1 / sum(kappa1)
kappa2 <- c(4,1,1,4)
kappa2 <- kappa2 / sum(kappa2)
kappa3 <- c(0,0,1,1)
kappa3 <- kappa3 / sum(kappa3)
kappa4 <- c(1,1,0,0)
kappa4 <- kappa4 / sum(kappa4)

x0 <- seq(0, 1, length=1001)
h0_basis <- create.bspline.basis(c(0, 1), 5)
h0_basis_mat <- eval.basis(h0_basis, x0)
h1 <- h0_basis_mat \%*\% cumsum(c(0, kappa0))
plot(h1~x0, type="l")

dat0 <- sim_warping_mixture(100, rep(1/3, 3),
                    rbind(kappa1,
                          kappa2,
                          kappa3),
                    ni = 201,
                    tau = 30,
                    mu_sh = -25, mu_sc = 500,
                    sd_sh = 10, sd_sc=50, sd_err = 5)

ggplot(dat0) +
  geom_line(aes(x=x, y=y, group=id, col=clust))
ggplot(dat0) +
  geom_line(aes(x=x, y=warped_x, group=id, col=clust))

}
